Safe write tags
Ensure that the tag is readable before returning. This works around a
problem where there is an unknown delay after setting a tag on a blob
before you can query for the blob by the tag. When we return
immediately, this causes bugs because Porter expects to be able to query
for the data immediately using the crud.Store interface which relies on
the tags.

So after setting a tag on a blob, query for it in a loop until we get
the blob back, then allow the call to return.