Cleanup secret name before use
Azure Key Vault only allows a limited set of characters in secret names, alphanumeric with hyphens with 127 characters max. I have updated the plugin to replace any invalid characters with a hyphen before storing or retrieving. When too long, I keep as much of the key as possible and then append the md5 sum of the original key to the end.

I've also added more logging so that we can track down when the value was cleaned and what the originally requested key was.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>