Trace azure plugin to open telemetry as separate service
When we emit traces to open telemetry from the azure plugin, change the service name to "porter.plugins.azure" and under its own root span (for distributed tracing). Right now we see the traces from the plugin, but it's not easy to isolate just stuff from a particular plugin since it's being grouped with the porter service.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>