Update to use Azure SDK modules
Autorest is legacy, see https://learn.microsoft.com/en-us/azure/developer/go/azure-sdk-authorization,
for more information.
The result is that the custom credentials functions is no longer needed,
simplifying the code. A downside is that this custom prefixes are no
easily supported, but Microsoft have standardized everything be prefixed
with `AZURE_`, and the possibility to changed the prefix was not
documented.

Signed-off-by: Kim Christensen <kmch@simcorp.com>